+++ BBBWebApiCodeFirst/DataTransferObjects/ByDayByPeriodByActivityDTO.cs
@@ -7,7 +7,7 @@ namespace BBBWebApiCodeFirst.DataTransferObjects
 {
     public class ByDayByPeriodByActivityDTO
     {
-        public int IdDay { get; set; }
+        public double IdDay { get; set; }
 
         public string Day { get; set; }
