+++ BBBWebApiCodeFirst/DataTransferObjects/ByWeekdaysByPeriodByActivityDTO.cs
@@ -7,7 +7,7 @@ namespace BBBWebApiCodeFirst.DataTransferObjects
 {
     public class ByWeekdaysByPeriodByActivityDTO
     {
-        public int IdDay { get; set; }
+        public double IdDay { get; set; }
 
         public string Day { get; set; }
