+++ BBBWebApiCodeFirst/DataTransferObjects/WeekendByPeriodByActivityDTO.cs
@@ -7,7 +7,7 @@ namespace BBBWebApiCodeFirst.DataTransferObjects
 {
     public class WeekendByPeriodByActivityDTO
     {
-        public int IdDay { get; set; }
+        public double IdDay { get; set; }
 
         public string Day { get; set; }
